Support topolgies with dynamic output topics (#47)
If a topology uses a TopicNameExtractor, iterating over the sub-topologies doesn't yield the topic name, as it is not known beforehand. Such cases would throw a NPE (see #46), because we added null to the set of output topics.

This PR just adds a check for these cases. If a TopicnameExtractor is used, developers must manually add the output topic to the set with testTopology.getOutputTopics().add("output").